# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/transaksi": {
            "get": {
                "description": "Get a list of Transaksi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Get all Transaksi.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaksi"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Transaksi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Create New Transaksi.",
                "parameters": [
                    {
                        "description": "the body to create a new Transaksi",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.transaksiInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaksi"
                        }
                    }
                }
            }
        },
        "/transaksi/{id}": {
            "get": {
                "description": "Get an Transaksi by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Get Transaksi.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaksi id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaksi"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Transaksi by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Delete one Transaksi.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaksi id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Transaksi by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Update Transaksi.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaksi id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age transaksi category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.transaksiInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaksi"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.transaksiInput": {
            "type": "object",
            "properties": {
                "biaya_dokter": {
                    "type": "string"
                },
                "biaya_kamar": {
                    "type": "string"
                },
                "biaya_obat": {
                    "type": "string"
                },
                "diagnosaID": {
                    "type": "integer"
                },
                "total": {
                    "type": "string"
                }
            }
        },
...
        "models.Transaksi": {
            "type": "object",
            "properties": {
                "biaya_dokter": {
                    "type": "string"
                },
                "biaya_kamar": {
                    "type": "string"
                },
                "biaya_obat": {
                    "type": "string"
                },
                "diagnosaID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "total": {
                    "type": "string"
                }
            }
        }
    }
}